refactor: mv connectors
