+++ src/app/(settings)/components/ConnectMPCWalletButton.tsx
@@ -7,7 +7,7 @@ import { connect, disconnect } from '@wagmi/core';
 import { useAsyncFn } from 'react-use';
 import { useAccount } from 'wagmi';
 
-import { createParticleConnector } from '@/app/connectors/ParticleConnector.js';
+import { createParticleConnector } from '@/connectors/ParticleConnector.js';
 import LoadingIcon from '@/assets/loading.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
 import { config } from '@/configs/wagmiClient.js';
