+++ src/connectors/ParticleConnector.ts
